+++ src/basic/models/product.ts
@@ -1,0 +1,2 @@
+
+
